+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+ - Add name to routes and update config file.
+
+### Added
  - Add `logout` method for web on `auth:classic` module.
  - Add `start` and `check` methods for web on `authorize:classic` module.
  - Add `logout_route`, `routes.api`, and `routes.web` to config file.